feat: CRUD de veículos completo e validado via Postman; correção de tipos decimais e segurança liberada para integração

+++ src/main/java/br/com/edu/fatec/IPEMControl/Repository/VeiculoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
+public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
     // O Spring Data JPA já cria os métodos save, find, delete etc. automaticamente
 }
